+++ braindecode/models/__init__.py
@@ -9,6 +9,9 @@ from .hybrid import HybridNet
 from .shallow_fbcsp import ShallowFBCSPNet
 from .eegresnet import EEGResNet
 from .eeginception import EEGInception
+from .eeginception_erp import EEGInceptionERP
+from .eeginception_mi import EEGInceptionMI
+from .atcnet import ATCNet
 from .tcn import TCN
 from .sleep_stager_chambon_2018 import SleepStagerChambon2018
 from .sleep_stager_blanco_2020 import SleepStagerBlanco2020
@@ -17,4 +20,3 @@ from .tidnet import TIDNet
 from .usleep import USleep
 from .util import get_output_shape, to_dense_prediction_model
 from .modules import TimeDistributed
-