+++ braindecode/models/modules.py
@@ -4,6 +4,7 @@
 
 import numpy as np
 
+import torch
 from torch import nn
 import torch.nn.functional as F
 
@@ -189,3 +190,111 @@ class TimeDistributed(nn.Module):
         b, s, c, t = x.shape
         out = self.module(x.view(b * s, c, t))
         return out.view(b, s, -1)
+
+
+class CausalConv1d(nn.Conv1d):
+    """Causal 1-dimensional convolution
+
+    Code modified from [1]_ and [2]_.
+
+    Parameters
+    ----------
+    in_channels : int
+        Input channels.
+    out_channels : int
+        Output channels (number of filters).
+    kernel_size : int
+        Kernel size.
+    dilation : int, optional
+        Dilation (number of elements to skip within kernel multiplication).
+        Default to 1.
+    **kwargs :
+        Other keyword arguments to pass to torch.nn.Conv1d, except for
+        `padding`!!
+
+    References
+    ----------
+    .. [1] https://discuss.pytorch.org/t/causal-convolution/3456/4
+    .. [2] https://gist.github.com/paultsw/7a9d6e3ce7b70e9e2c61bc9287addefc
+    """
+    def __init__(
+        self,
+        in_channels,
+        out_channels,
+        kernel_size,
+        dilation=1,
+        **kwargs,
+    ):
+        assert "padding" not in kwargs, (
+            "The padding parameter is controlled internally by "
+            f"{type(self).__name__} class. You should not try to override this"
+            " parameter."
+        )
+
+        super().__init__(
+            in_channels=in_channels,
+            out_channels=out_channels,
+            kernel_size=kernel_size,
+            dilation=dilation,
+            padding=(kernel_size - 1) * dilation,
+            **kwargs
+        )
+
+    def forward(self, X):
+        out = super().forward(X)
+        return out[..., :-self.padding[0]]
+
+
+class MaxNormLinear(nn.Linear):
+    """Linear layer with MaxNorm constraining on weights.
+
+    Equivalent of Keras tf.keras.Dense(..., kernel_constraint=max_norm())
+    [1, 2]_. Implemented as advised in [3]_.
+
+    Parameters
+    ----------
+    in_features: int
+        Size of each input sample.
+    out_features: int
+        Size of each output sample.
+    bias: bool, optional
+        If set to ``False``, the layer will not learn an additive bias.
+        Default: ``True``.
+
+    References
+    ----------
+    .. [1] https://keras.io/api/layers/core_layers/dense/#dense-class
+    .. [2] https://www.tensorflow.org/api_docs/python/tf/keras/constraints/
+           MaxNorm
+    .. [3] https://discuss.pytorch.org/t/how-to-correctly-implement-in-place-
+           max-norm-constraint/96769
+    """
+    def __init__(
+        self,
+        in_features,
+        out_features,
+        bias=True,
+        max_norm_val=2,
+        eps=1e-5,
+        **kwargs
+    ):
+        super().__init__(
+            in_features=in_features,
+            out_features=out_features,
+            bias=bias,
+            **kwargs
+        )
+        self._max_norm_val = max_norm_val
+        self._eps = eps
+
+    def forward(self, X):
+        self._max_norm()
+        return super().forward(X)
+
+    def _max_norm(self):
+        with torch.no_grad():
+            norm = self.weight.norm(2, dim=0, keepdim=True).clamp(
+                min=self._max_norm_val / 2
+            )
+            desired = torch.clamp(norm, max=self._max_norm_val)
+            self.weight *= (desired / (self._eps + norm))