+++ test/unit_tests/models/test_models.py
@@ -13,7 +13,8 @@ import pytest
 from braindecode.models import (
     Deep4Net, EEGNetv4, EEGNetv1, HybridNet, ShallowFBCSPNet, EEGResNet, TCN,
     SleepStagerChambon2018, SleepStagerBlanco2020, SleepStagerEldele2021, USleep,
-    DeepSleepNet, EEGITNet, EEGInception, TIDNet)
+    DeepSleepNet, EEGITNet, EEGInception, EEGInceptionERP, EEGInceptionMI, TIDNet, ATCNet)
+
 
 from braindecode.util import set_random_seeds
 
@@ -42,7 +43,8 @@ def check_forward_pass(model, input_sizes, only_check_until_dim=None):
     assert y_pred_new.shape[:only_check_until_dim] == (
         input_sizes['n_samples'], input_sizes['n_classes'])
     np.testing.assert_allclose(y_pred.detach().cpu().numpy(),
-                               y_pred_new.detach().cpu().numpy())
+                               y_pred_new.detach().cpu().numpy(),
+                               atol=1e-4, rtol=0)
 
 
 def test_shallow_fbcsp_net(input_sizes):
@@ -121,8 +123,9 @@ def test_eegitnet(input_sizes):
     check_forward_pass(model, input_sizes,)
 
 
-def test_eeginception(input_sizes):
-    model = EEGInception(
+@pytest.mark.parametrize("model_cls", [EEGInception, EEGInceptionERP])
+def test_eeginception_erp(input_sizes, model_cls):
+    model = model_cls(
         n_classes=input_sizes['n_classes'],
         in_channels=input_sizes['n_channels'],
         input_window_samples=input_sizes['n_in_times'])
@@ -130,11 +133,12 @@ def test_eeginception(input_sizes):
     check_forward_pass(model, input_sizes,)
 
 
-def test_eeginception_n_params():
+@pytest.mark.parametrize("model_cls", [EEGInception, EEGInceptionERP])
+def test_eeginception_erp_n_params(model_cls):
     """Make sure the number of parameters is the same as in the paper when
     using the same architecture hyperparameters.
     """
-    model = EEGInception(
+    model = model_cls(
         in_channels=8,
         n_classes=2,
         input_window_samples=128,  # input_time
@@ -147,6 +151,89 @@ def test_eeginception_n_params():
 
     n_params = sum(p.numel() for p in model.parameters() if p.requires_grad)
     assert n_params == 14926  # From paper's TABLE IV EEG-Inception Architecture Details
+
+
+def test_eeginception_mi(input_sizes):
+    sfreq = 250
+    model = EEGInceptionMI(
+        n_classes=input_sizes['n_classes'],
+        in_channels=input_sizes['n_channels'],
+        input_window_s=input_sizes['n_in_times'] / sfreq,
+        sfreq=sfreq,
+    )
+
+    check_forward_pass(model, input_sizes,)
+
+
+@pytest.mark.parametrize(
+    "n_filter,reported",
+    [(6, 51386), (12, 204002), (16, 361986), (24, 812930), (64, 5767170)]
+)
+def test_eeginception_mi_binary_n_params(n_filter, reported):
+    """Make sure the number of parameters is the same as in the paper when
+    using the same architecture hyperparameters.
+
+    Note
+    ----
+    For some reason, we match the correct number of parameters for all
+    configurations in the binary classification case, but none for the 4-class
+    case... Should be investigated by contacting the authors.
+    """
+    model = EEGInceptionMI(
+        in_channels=3,
+        n_classes=2,
+        input_window_s=3.,  # input_time
+        sfreq=250,
+        n_convs=3,
+        n_filters=n_filter,
+        kernel_unit_s=0.1,
+    )
+
+    n_params = sum(p.numel() for p in model.parameters() if p.requires_grad)
+    # From first column of TABLE 2 in EEG-Inception paper
+    assert n_params == reported
+
+
+def test_atcnet(input_sizes):
+    sfreq = 250
+    input_sizes["n_in_times"] = 1125
+    model = ATCNet(
+        n_channels=input_sizes['n_channels'],
+        n_classes=input_sizes['n_classes'],
+        input_size_s=input_sizes['n_in_times'] / sfreq,
+        sfreq=sfreq,
+    )
+
+    check_forward_pass(model, input_sizes,)
+
+
+def test_atcnet_n_params():
+    """Make sure the number of parameters is the same as in the paper when
+    using the same architecture hyperparameters.
+    """
+    n_windows = 5
+    att_head_dim = 8
+    att_num_heads = 2
+
+    model = ATCNet(
+        n_channels=22,
+        n_classes=4,
+        input_size_s=4.5,
+        sfreq=250,
+        n_windows=n_windows,
+        att_head_dim=att_head_dim,
+        att_num_heads=att_num_heads,
+    )
+
+    n_params = sum(p.numel() for p in model.parameters() if p.requires_grad)
+
+    # The paper states the models has around "115.2 K" parameters in its
+    # conclusion. By analyzing the official tensorflow code, we found indeed
+    # 115,172 parameters, but these take into account untrainable batch norm
+    # params, while the number of trainable parameters is 113,732.
+    official_code_nparams = 113_732
+
+    assert n_params == official_code_nparams
 
 
 @pytest.mark.parametrize(